fix: stop rendering Auth0 components when env vars not provided to surpress unexpected redirect
